Serve static files
Module: 05 Working with Express.js
Video: 019 Serving Files Statically

To serve static files (access them through file system, not router or any middleware), we use static() method of the express object. We should pass in a path to this method, a folder  we want to use for static files.

Basically, we grant read access to a folder content (CSS, JS, images, etc).