Use Handlebars as templating engine
Although we installed the handlebars package, unlike Pug, we need to manually import it using require. The require function is used to load the 'express-handlebars' module so we can use it in our application.

Then we also should register Handlebars as the template engine for .hbs files (the extension is up to you).

We also should change the 'view engine' property to 'hbs' of course.

To fix errors we should use change expressHbs() to expressHbs.engine() and also we should pass layout: false if we don't use layout.