Handle products.json being empty or invalid json
When the file content is successfully read, we try to parse it as JSON. With the previous code, if the file content was empty or not valid JSON, the JSON.parse would throw an error.

Now in such cases the error will be caught by the catch block and we call the callback cb with an empty array.