Pass Data with POST request
Unlike GET, when doing POST request we can pass data in the request body.

So by submitting a form, using POST method, we can pass the form data through body.

So, here we don't need to put the Product ID into the URL, as we had to when sending GET requests.