Retrieve data from MySQL database
Connect our Node.js app to a MySQL database and log data in console.

To use SQL inside Node, we need to install mysql2 package first.

For the next step, we will create a pool of connections, using the "mysql" obeject's createPool property (in database.js). Then we export the pool, using the promise method.

In app.js, we execute a query on our pool, on the products table. The execute function is some so-called promise (A promise is a basic JS object).
We chain the functions then() and catch() to it.