Fetch Single product info
We use WHERE to return specific product.
Note that db.execute() method returns a promise.

If the promise resolved, it typically passes two arguments to the .then() method. First is the actual data and the second is the meta data of the fields.

So, here we used only the first argument of then(). We didn't write the second argument.